Switch from persistence.xml file for configuration
The META-INF resource folder was not visable to the hibernate and
java persistance frameworks when the project had been compiled into
a run time image. This meant that the java persistance framework was
unable to access its configuration file.

To solve the issue, configuration was switched from the persistence.xml
file to the SQLiteProvider class with contains all of the configuration
information the xml file contained. This change also required a minor
change to how the DatabaseManager gets the EntityManagerFactory.

Next, the create tables sql script was moved from the META-INF folder
into an sql folder so that it can be accessed when hibernate performs
its start up actions.