Add creation of application's runtime image
Creation of the runtime image requires all dependencies to be modules.
However, since hibernate is not a modular dependency, creating the
runtime image requried adding a module info descriptor to its jar
and the jars of its dependencies during the build process.

The maven dependency plugin has been added to copy all of the runtime
dependencies into the build folder so that they can easily be added to
the module path during runtime image creation.

The moditech plugin was added to the project to perform the
insertion of the module descriptor and the creation of the runtime
image. A series of module-info.java files have been added for each
non-modular dependency and the moditech plugin inserts these into the
jar files of the non-modular dependencies, making them modular. The
plugin then creates the runtime image from both the existing
dependencies and the newly modular dependencies.

The created runtime image currently only works on linux and work will
have to be done to create a version that works on windows.