Fix subitem not displaying bug
This bug was caused by the javafx.base module being unable to access
the properties of the subitem. This was because the entitiy package
was not opened to this module, so the modular system was dening it
access when it was trying to populate the subitem table. Soultion
was to open access of the entitiy package to the javafx.base module.