Add automated test workflow
This workflow will run all of the tests in the project and will also
build the project. This is so that the project tests can be run
frequencly and that errors with the build process can be caught before
a release is made.

This work flow will run in three seperate instances. First, it will run
if a push is made to a non-master branch. Second, it will run when a
pull request is made to the master branch. Third, it will run on the
master branch everyday at mid-day.