Change build to create installier
The build process of the application has been changed from creating
a runtime image of the application to creating a system dependent
installier package. This was done by changed from using jlink in the
build to using jpackage.

As the produced installer package is system dependent, The build will
need to be performed on the target system. Build profiles have been
set up so that system specific jpackage options will be suppliered to
the plugin when the build is on the correct type of system.